feat: 后台管理系统

+++ springboot/system/src/main/java/org/example/springboot/system/domain/entity/CountView.java
@@ -43,7 +43,7 @@ public class CountView extends BaseEntity {
      * 业务类型
      */
     @Schema(description = "业务类型")
-    private Byte bizType;
+    private Integer bizType;
     /**
      * 浏览
      */
